Even More Filters
I have found even more subhuman garbage on YouTube, what a surprise. I blocked as much as I could. One of the filters in here may have adverse affects, it has 'slasher' in it and it may block <em>anything</em> with the term 'slasher' in it. Let me know if any of these filters have negative effects. On a side note, I am aware that if you directly attempt to "test" these filters by searching for Reddit reading videos, your browser will slow down a lot as your adblocker repeatedly deletes 90%+ of the page and YouTube tries to refill it. Enjoy!