New Spam Domain
Received a spam text containing the new domain. It claimed to be for tracking an Amazon purchase. I'm partially glad the last owner of my phone number was dumb as dirt and signed up for BS like this because I can expand this filter list using the spam messages that I get. Still annoying to deal with but it's whatever.