COVID-19 Scam Domain
This domain was used for by a malicious .xls file that was spread via email regarding COVID-19. This acted as a home server that the victim's PC would connect to. While this filter will not protect against the attack if the file is run, it will prevent access to the site.