Big Scam Update
Added a lot of scam domains peddled by hijacked YouTube channels. The rabbit hole surrounding this is massive and I hope to expand this list further. Note that most of these scams are already blocked by these lists as all of the sites typically use the same "human verification" site, which this list blocks.