Create TH-Twitch-Filters
This is a somewhat intensive filter list that removes some, not all, of the garbage found in Twitch chat. For example, some people spam emotes, so my list blocks all chat containing any emotes whatsoever. This list is by no means perfect but helps to take out some of the trash.